postgres now working

+++ services/appointments-service/src/models/service.rs
@@ -1,10 +1,12 @@
 use serde::Serialize;
+use sqlx::{Decode, FromRow};
+use bigdecimal::BigDecimal;
 
-#[derive(Serialize)]
+#[derive(FromRow, Debug, Serialize)]
 pub struct Service{
-    pub id:u32,
+    pub id:i32,
     pub name:String,
-    pub duration_minutes:u32,
-    pub price:f32
+    pub duration_minutes:i32,
+    pub price:BigDecimal
 }
